requirements.txt
tensorflow==2.10.0
tensorflow-directml-plugin
numpy==1.26.4
protobuf==3.19.6
scipy
scikit-learn
opencv-python-headless
streamlit
matplotlib
Pillow
tqdm
uvicorn
albumentations
keras
openai
python-multipart
fastapi
python-dotenv